move post and edit buttons back to FiresidePost
