+++ src/app/components/activity/feed/event-item/event-item.ts
@@ -28,7 +28,6 @@ import AppUserVerifiedTick from '../../../../../_common/user/verified-tick/verif
 import { Store } from '../../../../store';
 import AppEventItemControls from '../../../event-item/controls/controls.vue';
 import AppPollVoting from '../../../poll/voting/voting.vue';
-import { PostEditModal } from '../../../post/edit-modal/edit-modal-service';
 import AppActivityFeedCommentVideo from '../comment-video/comment-video.vue';
 import AppActivityFeedDevlogPostMedia from '../devlog-post/media/media.vue';
 import AppActivityFeedDevlogPostSketchfab from '../devlog-post/sketchfab/sketchfab.vue';
@@ -116,31 +115,6 @@ export default class AppActivityFeedEventItem extends Vue {
 		if (this.eventItem.type === EventItem.TYPE_POST_ADD) {
 			return this.eventItem.action as FiresidePost;
 		}
-	}
-
-	get canPublish() {
-		if (!this.post) {
-			return false;
-		}
-
-		return (
-			this.post.isDraft &&
-			!this.post.isScheduled &&
-			this.post.hasLead &&
-			this.post.canPublishToCommunities()
-		);
-	}
-
-	get hasPerms() {
-		if (!this.app.user || !this.post) {
-			return false;
-		}
-
-		return this.post.isEditableByUser(this.app.user);
-	}
-
-	get shouldShowEdit() {
-		return this.hasPerms;
 	}
 
 	get video() {
@@ -269,23 +243,6 @@ export default class AppActivityFeedEventItem extends Vue {
 	destroyed() {
 		this.feedComponent = undefined as any;
 		this.resizeSensor = undefined;
-	}
-
-	async openEdit() {
-		if (!this.post) {
-			return;
-		}
-
-		if (await PostEditModal.show(this.post)) {
-			this.onPostEdited(this.eventItem);
-		}
-	}
-
-	async publish() {
-		if (this.post) {
-			await this.post.$publish();
-			this.onPostPublished(this.eventItem);
-		}
 	}
 
 	/**
